pyproject.toml
[project]
name = "yolol-simulator"
version = "0.1.0"
description = "A Discord bot that simulates Yolol"
authors = [
    {name = "Ankur Bagchi", email = "your.email@example.com"},
]
dependencies = [
    "discord.py>=2.0.0",
    "python-dotenv>=0.19.0",
    "openai>=1.76.0",
]
requires-python = ">=3.8"
license = {text = "MIT"}


[tool.pdm]
distribution = false
